+++ controller/konnect/ops/ops_errors.go
@@ -76,6 +76,20 @@ func (e CantPerformOperationWithoutControlPlaneIDError) Error() string {
 	)
 }
 
+// CantPerformOperationWithoutNetworkIDError is an error indicating that an
+// operation cannot be performed without a Konnect network ID.
+type CantPerformOperationWithoutNetworkIDError struct {
+	Entity entity
+	Op     Op
+}
+
+func (e CantPerformOperationWithoutNetworkIDError) Error() string {
+	return fmt.Sprintf(
+		"can't %s %s %s without a Konnect Cloud Gateway Network ID",
+		e.Op, e.Entity.GetTypeName(), client.ObjectKeyFromObject(e.Entity),
+	)
+}
+
 type sdkErrorDetails struct {
 	TypeAt   string   `json:"@type"`
 	Type     string   `json:"type"`