+++ controller/konnect/reconciler_generic_rbac.go
@@ -12,6 +12,10 @@ package konnect
 //+kubebuilder:rbac:groups=konnect.konghq.com,resources=konnectcloudgatewaydataplanegroupconfigurations/status,verbs=update;patch
 //+kubebuilder:rbac:groups=konnect.konghq.com,resources=konnectcloudgatewaydataplanegroupconfigurations/finalizers,verbs=update;patch
 
+//+kubebuilder:rbac:groups=konnect.konghq.com,resources=konnectcloudgatewaytransitgateways,verbs=get;list;watch;update;patch
+//+kubebuilder:rbac:groups=konnect.konghq.com,resources=konnectcloudgatewaytransitgateways/status,verbs=update;patch
+//+kubebuilder:rbac:groups=konnect.konghq.com,resources=konnectcloudgatewaytransitgateways/finalizers,verbs=update;patch
+
 //+kubebuilder:rbac:groups=configuration.konghq.com,resources=kongconsumers,verbs=get;list;watch
 //+kubebuilder:rbac:groups=configuration.konghq.com,resources=kongconsumers/status,verbs=get;update;patch